- validate input

+++ src/shapes/IAreaCalculator.cs
@@ -2,5 +2,5 @@ namespace Shapes;
 
 public interface IAreaCalculator
 {
-    float CalculateArea();
+    double CalculateArea();
 }
